feat: Add script to configure Ubuntu machine
This commit adds a new script, `configure_machine.sh`, that can be used to configure an Ubuntu machine. The script checks if the OS is Ubuntu and if the `apt` package manager is installed. It then installs a list of required packages using `apt`. Additionally, it installs `zsh4humans` using either `curl` or `wget`.

The purpose of this script is to automate the setup process for an Ubuntu machine, making it easier for users to install the necessary packages and tools.