feat: Add script to stow all directories
This commit adds a new script, `stow_all.sh`, that can be used to stow all directories in the current directory. The script uses the `zsh` shell and iterates over all directories using a glob pattern. For each directory, it runs the `stow` command to symlink the contents of the directory to the specified target directory.

The purpose of this script is to automate the process of stowing directories, making it easier to manage and organize files in the home directory.